modified:   Practice/modular_inverse/modular_inverse/include/continued_fraction.h 	modified:   Practice/modular_inverse/modular_inverse/include/fermat.h

+++ Practice/modular_inverse/modular_inverse/include/fermat.h
@@ -1,13 +1,31 @@
 #ifndef FERMAT_H
 #define FERMAT_H
 
-// Проверка на простоту
+#include <cstdint> // Для int64_t
+
+/**
+ * Проверяет, является ли число простым.
+ * @param n Число для проверки.
+ * @return true, если число простое, иначе false.
+ */
 bool is_prime(int n);
 
-// Быстрое возведение в степень по модулю
-long long mod_exp(long long base, long long exponent, long long mod);
+/**
+ * Вычисляет base^exponent mod mod.
+ * @param base Основание.
+ * @param exponent Показатель степени.
+ * @param mod Модуль.
+ * @return Результат возведения в степень по модулю.
+ */
+int64_t mod_exp(int64_t base, int64_t exponent, int64_t mod);
 
-// Вычисление обратного элемента по теореме Ферма
-long long modInverseFermat(long long a, long long p);
+/**
+ * Находит обратный элемент по модулю с использованием теоремы Ферма.
+ * @param a Число, для которого ищется обратный элемент.
+ * @param p Простой модуль.
+ * @return Обратный элемент.
+ * @throws std::invalid_argument Если модуль не является простым числом или число кратно модулю.
+ */
+int64_t modInverseFermat(int64_t a, int64_t p);
 
 #endif
